MAXMEX
PROBLEM:
Given a sequence AA of NN integers, select the largest number of elements of AA such that the MEX of the chosen sequence of selected elements is MM, or determine that it’s impossible.

MEX of a sequence is the smallest positive integer not present in the sequence.

QUICK EXPLANATION
If the MEX of AA is already less than MM, it is impossible to make the MEX of any subset of AA exactly MM, hence it’s impossible to select a subset of AA with MEX MM
Otherwise, we can remove all occurrences of MM in AA. The remaining elements shall form the largest subset we can select which have MEX MM.
EXPLANATION
We can view selecting a subset of AA as the process of deleting the remaining elements. Let us assume the MEX of the sequence AA is XX.

Three cases arise

X = MX=M
The given array AA already has MEX MM, hence it is the largest subset which is the required answer.
X < MX<M
In this case, we actually need to add elements into AA in order to make MEX equal to MM. But, since we are not allowed to add new elements, it is impossible to make MEX MM.
X > MX>M
In this case, we know that all elements in the range [1, X-1][1,X−1] appear at least once and M \in [1, X-1]M∈[1,X−1]. If we delete all occurrences of MM from AA, the MEX of AA shall become MM, since all elements in the range [1, M-1][1,M−1] are still present, but MM is not. Hence, in this case, the required subset size is NN less the frequency of MM in AA.
Bonus
Given a sequence AA with NN elements, find the minimum number of operations to make MEX exactly MM if the following operations are allowed.

Add an element
Remove an element
Change the value of an element
TIME COMPLEXITY
The time complexity is O(N)O(N) or O(N*log(N))O(N∗log(N)) per test case, depending upon implementation.

SOLUTIONS: